Accurate type inference of Blocks
This patch attemps to accruately identify the type of block
(High/low/none). Lets first consider the following example

	let x = 5;
	let y = {
		// bunch of stuff
		x + 1
	}

It is clear that we can identify the type of solely based on
the type of the last statement which doesn't require inspection
of the full block. But that is not accuare, In the above case, the
type of "x + 1" does not depend on any other information from
inside the block, but considering something like:

	let y = {
		let x = 5;
		x + 1
	}

It becomes clearly obvios that by only inspecting the line "x + 1"
we reach the conclusion that "X" is of type none and "1" is of type
none. But that is not correct, we would need to inspect all the let
statements to identify that "x" is of type High.

Information about all let statements in a block is recordered when
processing thagt particular block but discarded afterwards. In principle
it should have been simple to maintain that information somewhere, but
in practice it is not quite simple. My idea is to store information about
inactive scopes(blocks) in a hashmap that is addressable by the block's
span. but proc_macro2 library had a different opinion. Span is neither
Hashable nor (Eq)-able. To be precise, Eq is implemented as unstable feature,
and it can only be enabled via some obscure rustc flags modifications, which
broke many thing. So it was not worth the effort anyways.

What I ended up using for identifying scopes is the debug formattable information
about the scopes, this is essentially the lowest form of hacks known to me in my
life... But I can either do that, or try to convince my profesor why I didn't do
it while puting my grades on the line.

Signed-off-by: Ahmed <ahmedsoliman@oddcoder.com>